user_group: Rename invite_user_group_picker_pill.ts to user_group_picker_pill.ts.
Fixes part of #32488.
