+++ tools/test-js-with-node
@@ -125,7 +125,6 @@ EXEMPT_FILES = make_set(
         "web/src/integration_url_modal.ts",
         "web/src/invite.ts",
         "web/src/invite_stream_picker_pill.ts",
-        "web/src/invite_user_group_picker_pill.ts",
         "web/src/left_sidebar_navigation_area.ts",
         "web/src/left_sidebar_navigation_area_popovers.ts",
         "web/src/lightbox.ts",
@@ -278,6 +277,7 @@ EXEMPT_FILES = make_set(
         "web/src/user_group_create_members_data.ts",
         "web/src/user_group_edit.ts",
         "web/src/user_group_edit_members.ts",
+        "web/src/user_group_picker_pill.ts",
         "web/src/user_group_popover.ts",
         "web/src/user_groups.ts",
         "web/src/user_pill.ts",
