+++ web/src/user_group_picker_pill.ts
